backend/requirements.txt
flask
sqlalchemy
flask-sqlalchemy
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
pytest
python-dotenv
